Reorganization to avoid circular import
