+++ pymc/progress_bar/rich_progress.py
@@ -32,7 +32,7 @@ from rich.style import Style
 from rich.table import Column, Table
 from rich.theme import Theme
 
-from pymc.progress_bar.progress import compute_draw_speed, default_progress_theme
+from pymc.progress_bar.utils import compute_draw_speed, default_progress_theme
 
 if TYPE_CHECKING:
     from rich.progress import TaskID
