+++ pymc/progress_bar/styles.py
